requirements.txt
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.6
idna==3.7
pycparser==2.22
pytz==2024.1
requests==2.31.0
urllib3==2.2.1
pyjwt==1.7.1